+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Sessions context with tests
 - Positions context with tests
 - Translations context with tests
+- uniquness for Example, Translation and Sentence
 
 ## [0.2.2] - 2019-01-20
 ### Added